types reorganize wip
