+++ packages/rhf-mui-components/src/common/types/components.ts
@@ -1,0 +1,17 @@
+export type OptionPrimitive = string | number;
+
+/**
+ * Generic type representing the possible value types for options
+ * used in form components like select, radio group, checkbox group, etc.
+ */
+export type OptionValue<
+  Option,
+  ValueKey extends string | undefined
+>
+  = Option extends OptionPrimitive
+    ? Option
+    : ValueKey extends keyof Option
+      ? Option[ValueKey] extends OptionPrimitive
+        ? Option[ValueKey]
+        : never
+      : never;
