+++ packages/rhf-mui-components/src/utils/object.ts
@@ -1,5 +1,6 @@
 import { type ReactNode } from 'react';
-import type { StringOrNumber, OptionPrimitive, OptionValue, StrNumObjOption } from '@/types';
+import type { OptionPrimitive, OptionValue } from '@/common';
+import type { StringOrNumber, StrNumObjOption } from '@/types';
 
 export function isKeyValueOption<LK extends string, VK extends string>(
   option: unknown,
